chore: refactored eslint configs and implemented
