+++ apps/docs/prettier.config.js
@@ -1,0 +1,6 @@
+import{ defaultConfig } from "@repo/prettier-config";
+
+export default {
+  ...defaultConfig,
+  "ignorePatterns": [".next/**", "node-modules"]
+}
