+++ packages/epic-postinstall/package.json
@@ -2,22 +2,43 @@
   "name": "epic-postinstall",
   "version": "0.0.1",
   "description": "A utility package for post-installation and cleanup scripts.",
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
+  "type": "module",
+  "main": "./dist/cjs/index.cjs",
+  "types": "./dist/esm/index.d.ts",
+  "exports": {
+    ".": {
+      "import": {
+        "types": "./dist/esm/index.d.ts",
+        "default": "./dist/esm/index.mjs"
+      },
+      "require": {
+        "types": "./dist/cjs/index.d.cts",
+        "default": "./dist/cjs/index.cjs"
+      }
+    }
+  },
   "scripts": {
     "clean": "rimraf dist node_modules",
     "postinstall": "echo \"postinstall goes here\" && exit 0",
-    "build": "tsc",
+    "build:esm": "tsc -p tsconfig.esm.json",
+    "build:cjs": "tsc -p tsconfig.cjs.json",
+    "build": "pnpm run build:esm && pnpm run build:cjs",
+    "watch:esm": "tsc -p tsconfig.esm.json --watch",
+    "watch:cjs": "tsc -p tsconfig.cjs.json --watch",
+    "watch": "pnpm run watch:esm & pnpm run watch:cjs",
     "lint": "eslint . --ext .ts",
-    "start": "node dist/index.js",
-    "run": "node dist/index.js",
-    "uninstall": "node dist/index.js --uninstall"
+    "start": "node dist/esm/index.mjs",
+    "run": "node dist/esm/index.mjs",
+    "uninstall": "node dist/esm/index.mjs --uninstall"
   },
   "bin": {
-    "epic-postinstall": "dist/index.js"
+    "epic-postinstall": "./dist/esm/index.mjs"
   },
   "devDependencies": {
-    "@types/node": "^22.15.24",
+    "@repo/typescript-config": "workspace:*",
+    "@repo/eslint-config": "workspace:*",
+    "@repo/prettier-config": "workspace:*",
+    "@types/node": "^22.15.29",
     "prettier": "^3.5.3",
     "rimraf": "^6.0.1",
     "typescript": "^5.8.3"
