+++ packages/eslint-config/index.js
@@ -1,7 +1,7 @@
 import { config } from "./base.js";
 import { nextJsConfig } from "./next.js";
 import { nodeJsConfig } from "./node-js.js";
-import { reactPackageConfig } from "./react-internal.js";
+import { reactPackageConfig } from "./react-Package.js";
 
 /** @type {import("eslint").Linter.Config[]} */
 export default config;
