+++ packages/ui/package.json
@@ -11,22 +11,23 @@
     "generate:component": "turbo gen react-component",
     "test": "echo \"tests: to be implemented\" && exit 0",
     "check-types": "tsc --noEmit",
-    "clean": "rimraf node_modules .turbo"
+    "clean": "rimraf node_modules .turbo dist"
   },
-  
   "dependencies": {
-    "react": "^19.0.0",
-    "react-dom": "^19.0.0"
+    "react": "^19.1.0",
+    "react-dom": "^19.1.0"
   },
   "devDependencies": {
     "@repo/eslint-config": "workspace:*",
     "@repo/typescript-config": "workspace:*",
-    "@turbo/gen": "^2.4.0",
+    "@repo/prettier-config" : "workspace:*", 
+    "@turbo/gen": "^2.5.4",
     "@types/node": "^22.13.0",
     "@types/react": "19.0.8",
     "@types/react-dom": "19.0.3",
     "eslint": "^9.21.0",
-    "typescript": "5.7.3",
-    "rimraf": "^6.0.1"
+    "prettier": "^3.5.3",
+    "rimraf": "^6.0.1",
+    "typescript": "5.7.3"
   }
 }
